% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_news.R
\name{check_news}
\alias{check_news}
\title{Check NEWS format}
\usage{
check_news()
}
\description{
Checks that NEWS.md exists and title lines follow the correct format.
\itemize{
\item If RELEASE: FALSE in the .github.yml, the test will be skipped. Useful when developing notes.
Otherwise
\item NEWS.md should exist in the root directory
\item Line 1 should be: # con$running
\item Line 2 : ## Version con$version \emph{20YY-MM-DD}
\item Line 3:  \verb{ * }
The
}
}
