% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkgtitle.R
\name{check_pkgtitle}
\alias{check_pkgtitle}
\title{Check PKG and Notes title}
\usage{
check_pkgtitle()
}
\description{
Package title should start with "Jumping Rivers: ".
Notes and Package titles should then be the same (excluding "Jumping Rivers : ").
}
