% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{make_wordlist}
\alias{make_wordlist}
\alias{check_spelling}
\title{Spell checkers}
\usage{
make_wordlist(spelling_results)

check_spelling()
}
\arguments{
\item{spelling_results}{A data frame - output of check spelling}
}
\description{
\code{check_spelling} runs a spell check on all c*.Rmd files.
}
\details{
\code{make_wordlist} takes the output of \code{check_spelling} and
creates a local \code{WORDLIST} file with exceptions. This should be
added to jrNotes.
}
