% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_template.R
\name{check_template}
\alias{check_template}
\title{Checks notes files are the same as template

Ensures that certain notes files are the same as the notes template.
These cannonical files are identical for all notes, e.g. Makefiles,
main.Rmd.

Typically this used by the gitlab-ci runner.}
\usage{
check_template(template_repo_loc = NULL)
}
\arguments{
\item{template_repo_loc}{Path to gitlab template location. If \code{NULL}
clones repo in \code{tempdir()}.}
}
\description{
Checks notes files are the same as template

Ensures that certain notes files are the same as the notes template.
These cannonical files are identical for all notes, e.g. Makefiles,
main.Rmd.

Typically this used by the gitlab-ci runner.
}
