% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_git_template.R
\name{clone_git_template}
\alias{clone_git_template}
\title{Create Notes}
\usage{
clone_git_template(name = NULL, path = NULL, push = FALSE)
}
\arguments{
\item{name}{The title of the notes directory you want to create
i.e. jRintroduction_notes. This defaults to "jRexample_notes".
To force you to create a directory with a name other than the default,
the default will break the function.
Unless the proper notation for a jumping rivers notes directory is
followed the function will have a paddy! i.e. must start with a "jr" and end in "_notes".}

\item{path}{The path where you want to place the notes. Default is NULL, which will create
the notes directory inside your current working directory.}

\item{push}{Default \code{FALSE}. TRUE to push and create the repo.}
}
\description{
Creates a Jumping Rivers style notes directory.
This assumes you have set up a private SSH key for Git.
}
