% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{collapse_news}
\alias{collapse_news}
\title{Collapse minor versions of news}
\usage{
collapse_news()
}
\description{
Collapse minor versions of news into one list
To use, open NEWS.md, highlight the sections that you want to collapse, and click the addin.
}
