% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_copy.R
\name{create_final}
\alias{create_final}
\title{Create copy of notes and practicals}
\usage{
create_final()
}
\description{
Renames main.pdf to notes_pkg.pdf. Also retrieves
the practials from the package and combines them.
}
