% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_live.R
\name{create_live_python}
\alias{create_live_python}
\title{Create live python}
\usage{
create_live_python(live_path)
}
\arguments{
\item{live_path}{The location to the live directory}
}
\description{
Creates a vm_scripts folder
Creates exercises.ipynb and solutions.ipynb from master_exercises.Rmd
Creates tutor.ipynb and a blank one from master_tutor.Rmd
Zips VM folder ready to be exposed as an artifact
}
