% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_live.R
\name{create_live_scripts}
\alias{create_live_scripts}
\title{Create scripts for online training}
\usage{
create_live_scripts(live_path = "../live")
}
\arguments{
\item{live_path}{The path to the live directory}
}
\description{
Creates a vm_scripts folder
Creates exercises.R and solutions.R from master_exercises.R
Creates tutor.R and tutor_blank.R from master_tutor.R
Zips VM folder ready to be exposed as an artifact
}
