% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_githook.R, R/get_logo_path.R
\name{create_githook}
\alias{create_githook}
\alias{create_logo}
\alias{create_title_page}
\alias{create_jrStyle}
\alias{create_version}
\title{Latex style and title pages}
\usage{
create_githook()

create_logo()

create_title_page(main = NULL, running = NULL, rss = NULL)

create_jrStyle()

create_version()
}
\arguments{
\item{main}{Front page title}

\item{running}{Running title}

\item{rss}{RSS boolean}
}
\description{
Returns path to logo
}
