% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jr_packages.R
\name{get_jr_packages}
\alias{get_jr_packages}
\title{Download R packages from jr-packages}
\usage{
get_jr_packages(path = ".")
}
\arguments{
\item{path}{Default \code{.}}
}
\description{
A convience function for downloading all jr-packages.
}
