% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo_path.R
\name{get_logo_path}
\alias{get_logo_path}
\alias{create_title_page}
\alias{create_jrStyle}
\title{Latex style and title pages}
\usage{
get_logo_path()

create_title_page(main, running)

create_jrStyle()
}
\arguments{
\item{main}{Front page title}

\item{running}{Running title}
}
\description{
Returns path to logo
}
