% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_githook.R, R/get_logo_path.R
\name{create_githook}
\alias{create_githook}
\alias{get_logos}
\alias{create_title_page}
\alias{create_jrStyle}
\alias{create_version}
\title{Create and get helper files}
\usage{
create_githook()

get_logos()

create_title_page(main = NULL, running = NULL, rss = NULL)

create_jrStyle()

create_version()
}
\arguments{
\item{main}{Front page title}

\item{running}{Running title}

\item{rss}{RSS boolean}
}
\description{
Get logos, sty files and title pages
}
