% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{has_version_been_updated}
\alias{has_version_been_updated}
\title{Config test}
\usage{
has_version_been_updated()
}
\description{
When the notes are updated, the version should be updated as well.
This function looks at the commit log to check if the notes have been
updated, the config.yml should also be updated.
}
\details{
This function is used in the .gitlab runner
}
