% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_code_hooks.R
\name{knitr_colours}
\alias{knitr_colours}
\alias{knitr_formatting}
\title{knitr code hooks}
\usage{
knitr_colours()

knitr_formatting()
}
\description{
These functions were hacked by Jamie. No-one really know what they do, or how
they work. In fact, I'm not convinced that knitr_formatting is correct. What's
color1 and color2????
}
