% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xelatex.R
\name{latexmk}
\alias{latexmk}
\title{Call latexmk}
\usage{
latexmk(fname)
}
\arguments{
\item{fname}{Latex file
Detects if the references.bib file contains entries and calls
bibtex appropriately.}
}
\description{
Call latexmk
}
