% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{live_comment_highlight}
\alias{live_comment_highlight}
\title{Inserts \code{#> }  in front of highlighted text.}
\usage{
live_comment_highlight()
}
\description{
Call this function as insert \code{#> } at the start of highlighted text.
}
