% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venv.R
\name{provision_venv}
\alias{provision_venv}
\title{Create and use a virtualenv with python packages listed in config.yml
installed}
\usage{
provision_venv()
}
\description{
Create and use a virtualenv with python packages listed in config.yml
installed
}
