% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_knitr_options}
\alias{set_knitr_options}
\title{Set consistent knitr options}
\usage{
set_knitr_options(tidy = FALSE, echo = FALSE, highlight = TRUE,
  comment = "#>", collapse = TRUE, cache = TRUE, fig.align = "center")
}
\arguments{
\item{tidy}{FALSE}

\item{echo}{FALSE}

\item{highlight}{TRUE}

\item{comment}{#>}

\item{collapse}{TRUE}

\item{cache}{TRUE}

\item{fig.align}{center}
}
\description{
Set consistent knitr options
}
