% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nice_par.R
\name{set_nice_par}
\alias{set_nice_par}
\title{Set nice par}
\usage{
set_nice_par(mar = c(3, 3, 2, 1), mgp = c(2, 0.4, 0), tck = -0.01,
  cex.axis = 0.9, las = 1, mfrow = c(1, 1), ...)
}
\arguments{
\item{mar}{c(3,3,2,1)}

\item{mgp}{c(2,0.4,0)}

\item{tck}{-.01}

\item{cex.axis}{0.9}

\item{las}{1}

\item{mfrow}{c(1, 1)}

\item{...}{Additional arguments passed to \code{par}}
}
\description{
Default parameters for plotting
}
