% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\title{Set Standard options}
\usage{
set_options(digits = 3, dplyr.print_min = 4, dplyr.print_max = 4,
  htmltools.dir.version = FALSE, min_pandoc_version = 2.2)
}
\arguments{
\item{digits}{3}

\item{dplyr.print_min}{4}

\item{dplyr.print_max}{4}

\item{htmltools.dir.version}{FALSE}

\item{min_pandoc_version}{2.2}
}
\description{
Set consistent options for notes / check pandoc version
}
