% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_palette.R
\name{set_palette}
\alias{set_palette}
\title{Set a consistent colour palette}
\usage{
set_palette(set = 0, alpha = 255)
}
\arguments{
\item{set}{default 0 (the standard R colour palette). \code{set=1} for
alternative palette}

\item{alpha}{Value between 0 and 255. Sets the transparancy level. Default 255.}
}
\description{
Set a consistent colour palette
}
