% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_news.R
\name{check_news}
\alias{check_news}
\title{Check NEWS format}
\usage{
check_news()
}
\description{
Checks that NEWS.md exists and title lines follow the correct format.
If the \code{pattern} is \code{NULL}, then the expected format is either:
\code{# pkg_name Version _YYYY-MM-DD_} or \code{# pkg_name _(development version)_}
}
