% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_template.R
\name{check_template}
\alias{check_template}
\title{Checks notes files are the same as template}
\usage{
check_template()
}
\description{
Ensures that certain notes files are the same as the notes template.
These canonical files are identical for all notes, e.g. Makefiles,
main.Rmd.
}
\details{
Typically this used by the gitlab-ci runner. The programming
language is inferred via \code{get_repo_language}.
}
