% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final.R
\name{get_python_pkg_name}
\alias{get_python_pkg_name}
\alias{get_r_pkg_name}
\alias{create_final}
\alias{create_final_python}
\title{Create copy of notes and practicals}
\usage{
get_python_pkg_name()

get_r_pkg_name()

create_final()

create_final_python()
}
\description{
Renames main.pdf to notes_pkg.pdf. Also retrieves
the practials from the package and combines them.
}
