% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_notes.R
\name{create_notes}
\alias{create_notes}
\title{Generate build script}
\usage{
create_notes(fnames = NULL, advert = TRUE, course_dep = TRUE)
}
\arguments{
\item{fnames}{If \code{NULL} scans for chaptersX.Rmd and appendix.Rmd.
Otherwise, just uses the names passed.}

\item{advert}{Should the advert be included. Default \code{TRUE}.}

\item{course_dep}{Should the course dependencies be included. Default \code{TRUE}.}
}
\description{
Scans for chaptersX.Rmd and appendix.Rmd and builds main.pdf
}
