% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_notes.R
\name{create_notes}
\alias{create_notes}
\title{Create Notes}
\usage{
create_notes(name = "jrNotes_Example", path = "~/")
}
\arguments{
\item{name}{The title of the notes directory you want to create
i.e. jRintroduction_notes. This defaults to "jrNotes_Example", which will break the function.
This is in order to stop you accidentally creating GitLab project after GitLab project!
Unless the proper notation for a jumping rivers notes directory is
followed the function will have a paddy!}

\item{path}{The path where you want to place the notes. Default is home.
This must be a full path from the home directory and end in a /.
However, the function will force you to do this.}
}
\description{
Creates a Jumping Rivers style notes directory.
This assumes you have set up a private SSH key for Git.
}
