% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hook.R, R/get_logo_path.R
\name{create_hook}
\alias{create_hook}
\alias{get_logo_path}
\alias{create_title_page}
\alias{create_jrStyle}
\alias{create_version}
\title{Latex style and title pages}
\usage{
create_hook()

get_logo_path()

create_title_page(main, running)

create_jrStyle()

create_version()
}
\arguments{
\item{main}{Front page title}

\item{running}{Running title}
}
\description{
Returns path to logo
}
