% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo_path.R
\name{get_logos}
\alias{get_logos}
\alias{create_title_page}
\alias{create_jrStyle}
\alias{create_version}
\title{Create and get helper files}
\usage{
get_logos()

create_title_page(main = NULL, running = NULL)

create_jrStyle()

create_version()
}
\arguments{
\item{main}{Front page title}

\item{running}{Running title}
}
\description{
Get logos, sty files and title pages
}
