% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_knitr_options.R
\name{set_knitr_options}
\alias{set_knitr_options}
\title{Set consistent knitr options}
\usage{
set_knitr_options(
  tidy = FALSE,
  echo = TRUE,
  highlight = TRUE,
  comment = "#>",
  collapse = TRUE,
  cache = TRUE,
  fig.align = "center",
  fig.width = 4,
  fig.height = 4,
  linewidth = 59,
  fig.retina = 1,
  out.width = "100\%",
  out.height = "100\%",
  dev = "cairo_pdf",
  dpi = 192,
  dev.args = list(png = list(type = "cairo-png")),
  ...
)
}
\arguments{
\item{tidy}{FALSE}

\item{echo}{TRUE}

\item{highlight}{TRUE}

\item{comment}{#>}

\item{collapse}{TRUE}

\item{cache}{TRUE}

\item{fig.align}{center}

\item{fig.width}{4}

\item{fig.height}{4}

\item{linewidth}{Code chunk line width - default 59}

\item{fig.retina}{Default 1}

\item{out.width}{Output width default 100\%}

\item{out.height}{Output height default 100\%}

\item{dev}{Default device \code{cairo_pdf}}

\item{dpi}{If changed to png, set to 192}

\item{dev.args}{If dev changed to png, use \code{cairo-png}}

\item{...}{Additional arguments passed to opt_chunk}
}
\description{
Set consistent knitr options
}
