% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_knitr_options}
\alias{set_knitr_options}
\title{Set consistent knitr options}
\usage{
set_knitr_options(tidy = FALSE, echo = TRUE, highlight = TRUE,
  comment = "#>", collapse = TRUE, cache = TRUE,
  fig.align = "center", fig.width = 4, fig.height = 4, ...)
}
\arguments{
\item{tidy}{FALSE}

\item{echo}{TRUE}

\item{highlight}{TRUE}

\item{comment}{#>}

\item{collapse}{TRUE}

\item{cache}{TRUE}

\item{fig.align}{center}

\item{fig.width}{4}

\item{fig.height}{4}

\item{...}{Additional arguments passed to opt_chunk}
}
\description{
Set consistent knitr options
}
